rasp/requirements.txt
RPi.GPIO==0.7.1
Flask==2.0.1
Werkzeug==2.0.1
requests==2.26.0
flask-cors==4.0.0